# MAINT: fix lobpcg performance regression