--- a/scipy/sparse/linalg/eigen/lobpcg/benchmarks/bench_lobpcg.py
+++ b/scipy/sparse/linalg/eigen/lobpcg/benchmarks/bench_lobpcg.py
@@ -46,7 +46,7 @@ def _as2d(ar):
     if ar.ndim == 2:
         return ar
     else:  # Assume 1!
-        aux = nm.array(ar, copy=False)
+        aux = np.array(ar, copy=False)
         aux.shape = (ar.shape[0], 1)
         return aux