Closes https://github.com/scipy/scipy/issues/4506

On the other hand, I think `lobpcg` is broken (https://github.com/scipy/scipy/issues/4174).

before:
```
               lobpcg benchmark sakurai et al.
==============================================================
      shape      | blocksize |    operation   |   time   
                                              | (seconds)
--------------------------------------------------------------
        (50, 50) |     3     |     lobpcg     |   0.99 
        (50, 50) |     3     |       eigh     |   0.01 
      (400, 400) |     3     |     lobpcg     |   9.98 
      (400, 400) |     3     |       eigh     |   0.10 
    (2400, 2400) |     3     |     lobpcg     |  18.56 
    (2400, 2400) |     3     |       eigh     |  14.83 
```

after:
```
                 lobpcg benchmark sakurai et al.
==============================================================
      shape      | blocksize |    operation   |   time   
                                              | (seconds)
--------------------------------------------------------------
        (50, 50) |     3     |     lobpcg     |   0.42 
        (50, 50) |     3     |       eigh     |   0.01 
      (400, 400) |     3     |     lobpcg     |   3.04 
      (400, 400) |     3     |       eigh     |   0.09 
    (2400, 2400) |     3     |     lobpcg     |   9.52 
    (2400, 2400) |     3     |       eigh     |  14.45 
```